Add files via upload
This project is about data analysis of Airline Passenger Satisfaction. This project aims to further improve the airline's services to it's customers by analyzing the rating of passengers' satisfaction. In this way, we can assure our loyal customers to continue availing the airline's services and invite new customers to shift in availing our services. As a data analyst, we will answer the question "What services of the airline can futher be improved to satisfy the needs of our current customers and to attract new customers to increase sales and profit?"